Rename wasm-pkg-loader to wasm-pkg-client
