+++ ci/run-ci.sh
@@ -4,4 +4,4 @@ set -ex
 
 cargo clippy --workspace --all-features
 cargo test --workspace --all-features
-(cd crates/wasm-pkg-loader/tests/e2e && cargo run)
+(cd crates/wasm-pkg-client/tests/e2e && cargo run)
