+++ crates/wasm-pkg-client/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "wasm-pkg-loader"
-description = "Wasm package loader"
-repository = "https://github.com/bytecodealliance/wasm-pkg-tools/tree/main/crates/wasm-pkg-loader"
+name = "wasm-pkg-client"
+description = "Wasm package client"
+repository = "https://github.com/bytecodealliance/wasm-pkg-tools/tree/main/crates/wasm-pkg-client"
 edition.workspace = true
 version.workspace = true
 authors.workspace = true
