+++ crates/wasm-pkg-client/src/local.rs
@@ -9,7 +9,7 @@ use tokio_util::io::ReaderStream;
 use wasm_pkg_common::{config::RegistryConfig, package::Version};
 
 use crate::{
-    source::{PackageSource, VersionInfo},
+    loader::{PackageLoader, VersionInfo},
     ContentDigest, Error, PackageRef, Release,
 };
 
@@ -18,14 +18,14 @@ pub struct LocalConfig {
     pub root: PathBuf,
 }
 
-/// A simple local filesystem-based PackageSource.
+/// A simple local filesystem-based backend.
 ///
 /// Each package release is a file: `<root>/<namespace>/<name>/<version>.wasm`
-pub struct LocalSource {
+pub struct LocalBackend {
     root: PathBuf,
 }
 
-impl LocalSource {
+impl LocalBackend {
     pub fn new(registry_config: RegistryConfig) -> Result<Self, Error> {
         let config = registry_config
             .backend_config::<LocalConfig>("local")?
@@ -47,7 +47,7 @@ impl LocalSource {
 }
 
 #[async_trait]
-impl PackageSource for LocalSource {
+impl PackageLoader for LocalBackend {
     async fn list_all_versions(&mut self, package: &PackageRef) -> Result<Vec<VersionInfo>, Error> {
         let mut versions = vec![];
         let package_dir = self.package_dir(package);
