+++ crates/wasm-pkg-client/tests/e2e/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "wasm-pkg-loader-e2e-tests"
+name = "wasm-pkg-client-e2e-tests"
 version = "0.1.0"
 edition = "2021"
 
@@ -13,6 +13,6 @@ oci-wasm = "0.0.3"
 tempfile = "3.9.0"
 tokio = { version = "1", default-features = false, features = ["fs"] }
 tokio-test = "0.4.3"
-wasm-pkg-loader = { path = "../.." }
+wasm-pkg-client = { path = "../.." }
 
 [workspace]
