+++ crates/wasm-pkg-client/tests/e2e/testdata/binary_wit.wasm
