+++ crates/wkg/Cargo.toml
@@ -15,5 +15,5 @@ tokio = { workspace = true, features = ["macros", "rt"] }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 wasm-pkg-common = { workspace = true }
-wasm-pkg-loader = { workspace = true }
+wasm-pkg-client = { workspace = true }
 wit-component = "0.207"
